feat(endpoint): add method delete() to the Doctor entity
This method allows the client to deactivate one record of the doctor table by the
id, with a DELETE request 'doctor/{id}'.

The implementation of this method does not delete the doctor permanently, just
sets the isActive boolean attribute to 'false'. In this way, the record can be
recovered, if needed. One new migration file was created to handle this update,
the new attribute, in the table.